feat(board): share the insights drill-in tiles with the read-only board
Extract the summary strip, Finance/Attendance/Enrolment/Performance/
Infrastructure tiles and their aggregate bar builders out of
app/(app)/insights/page.tsx into components/dashboard/insight-tiles.tsx,
imported 1:1 by BOTH /insights and /board (no duplication).

/board now reads the same getDirectorsInsights seam (dropping the direct
getSchoolRollup call) so it gains the class / year-group / subject
drill-downs; the shared PerformanceTile is a UNION keeping the board's
TERMINAL EXAMS/WASSCE section AND the drill-ins. The board keeps its own
framing (Board overview, read-only governance snapshot, /board/board-pack).

The Needs-attention derivation (buildAttention) moves to the server-only
seam; AttentionPanel takes a linkless flag so the board renders the same
signals as text rows with NO <a>/navigation (a board member is confined to
/board*), while /insights keeps its action links. Aggregate-only invariant
preserved and now covered by the shared-tiles source sweep + a render test.
