+++ omnischools/app/(board)/board/page.tsx
@@ -1,43 +1,28 @@
-import type { ReactNode } from "react";
 import { requireBoard } from "@/lib/auth/server";
-import {
-  getSchoolRollup,
-  type AttendanceArm,
-  type EnrolmentArm,
-  type InfrastructureSummary,
-  type NetPositionFinanceArm,
-  type PerformanceArm,
-  type RollupArm,
-  type TerminalResultsArm,
-  type TerminalResultSummary,
-} from "@/lib/rollup/school-rollup";
-import { boardGhs } from "@/lib/board/tiles";
+import { getDirectorsInsights, buildAttention } from "@/lib/insights/insights-data";
 import { ReportFilters } from "@/components/reports/report-filters";
-import { PerfBar } from "@/components/reports/report-kit";
 import {
-  TrendPill,
-  AbsencePanel,
-  SummaryCell,
-  Tile,
-  Line,
-  StreamCard,
-  StatusSplit,
-  FEMALE_HEX,
-  MALE_HEX,
-} from "@/components/board/board-tiles";
+  SummaryStrip,
+  AttentionPanel,
+  FinanceTile,
+  AttendanceTile,
+  EnrolmentTile,
+  PerformanceTile,
+  InfrastructureTile,
+} from "@/components/dashboard/insight-tiles";
 
 /**
- * GOV-4 · the read-only board/director dashboard (`/board`, `requireBoard()` — BOARD_MEMBER only).
- * Supersedes the GOV-2/3 minimal shell with Lucy's designed 5-tile dashboard: a summary scan strip +
- * Finance (full-width) · Attendance | Enrolment · Performance | Infrastructure. It reads the SESSION
- * school id (never a URL school id — R339); `?periodId` only picks the term. Every value flows through
- * the omit-not-fake convention:
- *   • treatment A — a solid-border reason for NOT_CAPTURED / NOT_APPLICABLE (no number);
- *   • treatment B — a real, coloured/typed zero for a captured 0;
- *   • treatment C — a dashed, italic "coming soon" for a not-yet-built PendingArm (never a number).
- * Finance appears exactly ONCE (the Finance tile owns the GHS streams; the summary strip carries
- * finance only as the collection-rate %). Trend pills come from EXPOSED deltas only — no fabricated
- * health verdict (the rollup strips ops thresholds — §10 honesty boundary).
+ * GOV-4 · the read-only board dashboard (`/board`, `requireBoard()` — BOARD_MEMBER only). It renders the
+ * SAME aggregate seam and SHARED tiles as Directors' Insights (`getDirectorsInsights` +
+ * `components/dashboard/insight-tiles.tsx`) — the summary scan strip, Finance (full-width), Attendance |
+ * Enrolment, Performance | Infrastructure, each with the class · year-group · subject drill-ins — but
+ * keeps the board's OWN framing: "Board overview", a "read-only governance snapshot", and the board-pack
+ * PDF under `/board` (so `requireBoard()`'s x-pathname confinement admits it).
+ *
+ * A board member is CONFINED to `/board*`, so the "Needs your attention" panel here is INFORMATIONAL:
+ * `linkless` renders each signal as a text row with NO navigation (a link to /billing etc. would be a
+ * dead end). Reads the SESSION school id (never a URL school id — R339); `?periodId` only picks the term.
+ * Everything is AGGREGATE — no student name / id / code / DOB (arms-length governance).
  */
 export const dynamic = "force-dynamic";
 
@@ -48,13 +33,13 @@ export default async function BoardPage({
 }) {
   const { school } = await requireBoard();
   const { periodId } = await searchParams;
-  const rollup = await getSchoolRollup(school.id, { periodId });
+  const data = await getDirectorsInsights(school.id, { periodId });
+  const { rollup, census } = data;
 
   const termLabel = rollup.period
     ? `${rollup.period.label} · ${rollup.period.academicYear}`
     : "No academic period configured";
-
-  const academic = academicSummary(rollup.performance);
+  const attention = buildAttention(data, termLabel);
 
   return (
     <div className="space-y-6">
@@ -86,526 +71,23 @@ export default async function BoardPage({
       />
 
       {/* ── Summary strip — the scan layer ── */}
-      <div className="grid grid-cols-2 gap-3 sm:grid-cols-3 lg:grid-cols-5">
-        <SummaryCell
-          lead
-          label="Students on roll"
-          value={
-            rollup.enrolment.status === "CAPTURED"
-              ? rollup.enrolment.data.roll.toLocaleString("en-GH")
-              : "—"
-          }
-          sub={
-            rollup.enrolment.status === "CAPTURED" ? (
-              <>
-                {rollup.enrolment.data.gender.male} boys · {rollup.enrolment.data.gender.female} girls{" "}
-                <TrendPill
-                  delta={rollup.enrolment.data.netChange}
-                  context="this term"
-                  flatLabel="no change"
-                />
-              </>
-            ) : (
-              rollup.enrolment.reason
-            )
-          }
-        />
+      <SummaryStrip rollup={rollup} />
 
-        <SummaryCell
-          label="Attendance rate"
-          value={
-            rollup.attendance.status === "CAPTURED"
-              ? rollup.attendance.data.schoolRate == null
-                ? "—"
-                : `${rollup.attendance.data.schoolRate}%`
-              : "—"
-          }
-          sub={
-            rollup.attendance.status === "CAPTURED" ? (
-              rollup.attendance.data.schoolDelta == null ? (
-                "(present + late) ÷ all marks"
-              ) : (
-                <TrendPill delta={rollup.attendance.data.schoolDelta} unit="pts" context="vs last term" />
-              )
-            ) : (
-              rollup.attendance.reason
-            )
-          }
-        />
-
-        <SummaryCell label="Academic standing" value={academic.value} sub={academic.sub} />
-
-        <SummaryCell
-          label="Fee collection"
-          value={
-            rollup.feeCollections.status === "CAPTURED"
-              ? `${rollup.feeCollections.data.collectionRate}%`
-              : "—"
-          }
-          sub={
-            rollup.feeCollections.status === "CAPTURED"
-              ? `${boardGhs(rollup.feeCollections.data.collected)} of ${boardGhs(rollup.feeCollections.data.billed)} billed`
-              : rollup.feeCollections.reason
-          }
-        />
-
-        {/* Cell 5 · Facilities — %-sound classrooms from the LATEST snapshot; honest reason when none. */}
-        <SummaryCell
-          label="Facilities"
-          value={
-            rollup.infrastructure.status === "CAPTURED"
-              ? rollup.infrastructure.data.classrooms.pctGood == null
-                ? "—"
-                : `${rollup.infrastructure.data.classrooms.pctGood}%`
-              : "—"
-          }
-          sub={
-            rollup.infrastructure.status === "CAPTURED"
-              ? `${rollup.infrastructure.data.classrooms.good}/${rollup.infrastructure.data.classrooms.total} classrooms sound`
-              : rollup.infrastructure.reason
-          }
-        />
-      </div>
+      {/* ── Needs your attention — INFORMATIONAL, link-free (the board is confined to /board*) ── */}
+      <AttentionPanel items={attention} linkless />
 
       {/* ── Detail tiles — the read layer ── */}
       <FinanceTile arm={rollup.netPositionFinance} />
 
       <div className="grid gap-4 lg:grid-cols-2">
-        <AttendanceTile arm={rollup.attendance} />
-        <EnrolmentTile arm={rollup.enrolment} />
+        <AttendanceTile arm={rollup.attendance} byLevel={data.attendanceByLevel} />
+        <EnrolmentTile arm={rollup.enrolment} census={census} />
       </div>
 
       <div className="grid gap-4 lg:grid-cols-2">
-        <PerformanceTile performance={rollup.performance} terminal={rollup.terminalResults} />
+        <PerformanceTile data={data} termLabel={termLabel} />
         <InfrastructureTile arm={rollup.infrastructure} />
       </div>
     </div>
   );
 }
-
-/* ───────────────────────────── Summary strip ───────────────────────────── */
-
-/** Cell 3 headline: prefer the captured Basic average; else Senior readiness; else the applicable
- *  tier's honest reason. NO blend across tiers (R357) — one tier's figure, never a mixed number. */
-function academicSummary(p: PerformanceArm): { value: string; sub: ReactNode } {
-  if (p.basic.status === "CAPTURED" && p.basic.data.overallAverage != null) {
-    const d = p.basic.data;
-    return {
-      value: `${d.overallAverage}%`,
-      sub: (
-        <>
-          {/* GOV-4a — pass rate on the standing cell; null renders absent, never "0%". */}
-          {d.passRate != null && <>{d.passRate}% pass · </>}
-          {d.gradedClasses} {d.gradedClasses === 1 ? "class" : "classes"} graded{" "}
-          <TrendPill delta={d.overallDelta} unit="pts" context="vs last term" />
-        </>
-      ),
-    };
-  }
-  if (p.senior.status === "CAPTURED") {
-    const d = p.senior.data;
-    return { value: `${d.subjectsReady}/${d.subjectsTotal}`, sub: "subjects STPSHS-ready" };
-  }
-  // Neither captured — surface the reason of the tier that APPLIES (skip a NOT_APPLICABLE tier).
-  const applicable = p.basic.status !== "NOT_APPLICABLE" ? p.basic : p.senior;
-  return { value: "—", sub: applicable.status === "CAPTURED" ? null : applicable.reason };
-}
-
-/* ───────────────────────────── Finance tile ───────────────────────────── */
-
-function FinanceTile({ arm }: { arm: RollupArm<NetPositionFinanceArm> }) {
-  return (
-    <Tile title="Financial" accent="position" className="col-span-full">
-      <p className="mt-1 max-w-2xl text-[13px] leading-relaxed text-navy-3">
-        Three separate records shown side by side. Fee collections and the school&apos;s books are kept
-        as separate ledgers and are not combined into a single profit; payroll is a current monthly
-        figure.
-      </p>
-
-      {arm.status !== "CAPTURED" ? (
-        <div className="mt-4">
-          <AbsencePanel>{arm.reason}</AbsencePanel>
-        </div>
-      ) : (
-        <div className="mt-4 grid gap-4 sm:grid-cols-3">
-          {/* Stream 1 · Fee collections — the feeCollections arm reused verbatim, + a rate bar. */}
-          <StreamCard title="Fee collections">
-            {arm.data.fees.status === "CAPTURED" ? (
-              <>
-                <Headline>{boardGhs(arm.data.fees.data.collected)}</Headline>
-                <Caption>collected · this term</Caption>
-                <div className="mt-3">
-                  <PerfBar value={arm.data.fees.data.collectionRate} tone="gold" suffix="%" />
-                </div>
-                <div className="mt-1.5 text-[11px] text-navy-3">
-                  {boardGhs(arm.data.fees.data.outstanding)} outstanding
-                </div>
-              </>
-            ) : (
-              <Reason>{arm.data.fees.reason}</Reason>
-            )}
-          </StreamCard>
-
-          {/* Stream 2 · Books — income / expense / net, all WITHIN the one books ledger. */}
-          <StreamCard title="Books (this term)">
-            {arm.data.books.status === "CAPTURED" ? (
-              <dl className="mt-2 space-y-1 text-[13px]">
-                <Line label="Income" value={boardGhs(arm.data.books.data.income)} />
-                <Line label="Expense" value={boardGhs(arm.data.books.data.expense)} />
-                <Line label="Net" value={boardGhs(arm.data.books.data.net)} strong />
-              </dl>
-            ) : (
-              <Reason>{arm.data.books.reason}</Reason>
-            )}
-          </StreamCard>
-
-          {/* Stream 3 · Payroll — school-paid gross monthly; GES/allowance are separate memos. */}
-          <StreamCard title="Payroll">
-            {arm.data.payroll.status === "CAPTURED" ? (
-              <>
-                <Headline>{boardGhs(arm.data.payroll.data.schoolPaidMonthlyTotal)}</Headline>
-                <Caption>school-paid · gross · monthly</Caption>
-                <div className="mt-2 text-[11px] leading-relaxed text-navy-3">
-                  GES-paid (memo, not added): {boardGhs(arm.data.payroll.data.gesPaidMonthlyMemo)}
-                  {arm.data.payroll.data.allowanceMonthlyMemo > 0 && (
-                    <>
-                      <br />
-                      Allowance (memo, not added):{" "}
-                      {boardGhs(arm.data.payroll.data.allowanceMonthlyMemo)}
-                    </>
-                  )}
-                </div>
-              </>
-            ) : (
-              <Reason>{arm.data.payroll.reason}</Reason>
-            )}
-          </StreamCard>
-        </div>
-      )}
-    </Tile>
-  );
-}
-
-/* ───────────────────────────── Attendance tile ───────────────────────────── */
-
-function AttendanceTile({ arm }: { arm: RollupArm<AttendanceArm> }) {
-  return (
-    <Tile
-      title="Attendance"
-      accent="this term"
-      meta={arm.status === "CAPTURED" ? `${arm.data.totalMarked.toLocaleString("en-GH")} marks recorded` : undefined}
-    >
-      {arm.status !== "CAPTURED" ? (
-        <div className="mt-4">
-          <AbsencePanel>{arm.reason}</AbsencePanel>
-        </div>
-      ) : (
-        <>
-          <div className="mt-3 flex items-center gap-3">
-            <div className="font-display text-3xl font-medium leading-none text-navy">
-              {arm.data.schoolRate == null ? "—" : `${arm.data.schoolRate}%`}
-            </div>
-            {arm.data.schoolDelta == null ? (
-              <span className="text-[11px] text-navy-3">(present + late) ÷ all marks</span>
-            ) : (
-              <TrendPill delta={arm.data.schoolDelta} unit="pts" context="vs last term" />
-            )}
-          </div>
-          <StatusSplit totals={arm.data.statusTotals} />
-        </>
-      )}
-    </Tile>
-  );
-}
-
-/* ───────────────────────────── Enrolment tile ───────────────────────────── */
-
-function EnrolmentTile({ arm }: { arm: RollupArm<EnrolmentArm> }) {
-  return (
-    <Tile
-      title="Enrolment"
-      accent="at a glance"
-      meta={arm.status === "CAPTURED" ? arm.data.levelSummary : undefined}
-    >
-      {arm.status !== "CAPTURED" ? (
-        <div className="mt-4">
-          <AbsencePanel>{arm.reason}</AbsencePanel>
-        </div>
-      ) : (
-        <EnrolmentBody d={arm.data} />
-      )}
-    </Tile>
-  );
-}
-
-function EnrolmentBody({ d }: { d: EnrolmentArm }) {
-  const dash = (n: number | null) => (n == null ? "—" : n.toLocaleString("en-GH"));
-  return (
-    <div className="mt-3 space-y-4">
-      {/* Roll headline + net-change pill. */}
-      <div className="flex items-center gap-3">
-        <div className="font-display text-3xl font-medium leading-none text-navy">
-          {d.roll.toLocaleString("en-GH")}
-        </div>
-        <TrendPill delta={d.netChange} context="this term" flatLabel="no change" />
-      </div>
-
-      {/* Gender mini-bar (school-stats pink/blue — a sanctioned non-token exception). */}
-      <div>
-        <div className="flex h-2.5 w-full overflow-hidden rounded-pill border border-border bg-bg">
-          {d.gender.female > 0 && (
-            <div style={{ flexGrow: d.gender.female, backgroundColor: FEMALE_HEX }} aria-hidden />
-          )}
-          {d.gender.male > 0 && (
-            <div style={{ flexGrow: d.gender.male, backgroundColor: MALE_HEX }} aria-hidden />
-          )}
-        </div>
-        <div className="mt-1.5 font-mono text-[10px] text-navy-3">
-          {d.gender.female}F · {d.gender.male}M
-        </div>
-      </div>
-
-      {/* Structure lines. */}
-      <dl className="space-y-1 text-[13px]">
-        <Line label="Active classes" value={d.activeClasses.toLocaleString("en-GH")} />
-        <Line label="Avg class size" value={d.avgClassSize.toLocaleString("en-GH")} />
-        <Line label="Teaching staff" value={d.teachingStaff.toLocaleString("en-GH")} />
-        <Line
-          label="Student : teacher"
-          value={d.studentTeacherRatio == null ? "—" : `${d.studentTeacherRatio}:1`}
-        />
-      </dl>
-
-      {/* Intake this term + lifetime exits. Term-windowed nulls render "—", never a fabricated 0. */}
-      <div className="border-t border-border pt-3">
-        <div className="text-[10px] font-bold uppercase tracking-[0.12em] text-navy-3">
-          Intake this term
-        </div>
-        <div className="mt-1 text-[13px] text-navy-2">
-          {dash(d.admissionsThisTerm)} new{" "}
-          <span className="text-navy-3">
-            ({dash(d.intakeFemale)}F · {dash(d.intakeMale)}M)
-          </span>
-        </div>
-        <div className="mt-2 text-[13px] text-navy-2">
-          <span className="text-navy-3">Lifetime exits:</span> {d.withdrew.toLocaleString("en-GH")}{" "}
-          withdrew · {d.transferred.toLocaleString("en-GH")} transferred ·{" "}
-          {d.graduated.toLocaleString("en-GH")} graduated ({d.lifetimeExits.toLocaleString("en-GH")}{" "}
-          total)
-        </div>
-        <p className="mt-1 text-[11px] leading-relaxed text-navy-3">
-          Withdrawals, transfers and graduations are current lifetime totals — per-term exit dating
-          arrives when status history is tracked.
-        </p>
-      </div>
-    </div>
-  );
-}
-
-/* ───────────────────────────── Performance tile ───────────────────────────── */
-
-function PerformanceTile({
-  performance,
-  terminal,
-}: {
-  performance: PerformanceArm;
-  terminal: TerminalResultsArm;
-}) {
-  const { basic, senior } = performance;
-  return (
-    <Tile title="Academic" accent="performance" meta="cross-tier · this term">
-      <div className="mt-3 space-y-3">
-        {/* Basic — omitted when the tier does not apply (omit-not-fake). */}
-        {basic.status !== "NOT_APPLICABLE" && (
-          <div>
-            <div className="text-[10px] font-bold uppercase tracking-[0.12em] text-navy-3">
-              Basic · gradebook
-            </div>
-            {basic.status === "CAPTURED" ? (
-              <div className="mt-1 flex flex-wrap items-center gap-x-3 gap-y-1">
-                <div className="font-display text-3xl font-medium leading-none text-navy">
-                  {basic.data.overallAverage == null ? "—" : `${basic.data.overallAverage}%`}
-                </div>
-                {/* GOV-4a — pass rate beside the average; null renders absent, never "0%". */}
-                {basic.data.passRate != null && (
-                  <span className="text-[11px] font-semibold text-navy-2">
-                    {basic.data.passRate}% pass rate
-                  </span>
-                )}
-                <span className="text-[11px] text-navy-3">
-                  {basic.data.gradedClasses} {basic.data.gradedClasses === 1 ? "class" : "classes"}{" "}
-                  graded
-                </span>
-                <TrendPill delta={basic.data.overallDelta} unit="pts" context="vs last term" />
-              </div>
-            ) : (
-              <p className="mt-1 text-[13px] leading-relaxed text-navy-3">{basic.reason}</p>
-            )}
-          </div>
-        )}
-
-        {/* Senior — STPSHS readiness completion counts (no scores, no names — §6.2). */}
-        {senior.status !== "NOT_APPLICABLE" && (
-          <div>
-            <div className="text-[10px] font-bold uppercase tracking-[0.12em] text-navy-3">
-              Senior · STPSHS readiness
-            </div>
-            {senior.status === "CAPTURED" ? (
-              <div className="mt-1 text-[13px] text-navy-2">
-                <span className="font-display text-xl font-medium text-navy">
-                  {senior.data.subjectsReady}
-                </span>{" "}
-                of {senior.data.subjectsTotal} subjects ready ·{" "}
-                <span className="font-semibold text-gold">{senior.data.subjectsPartial} partial</span>{" "}
-                · <span className="font-semibold text-terra">{senior.data.subjectsAtRisk} at risk</span>
-              </div>
-            ) : (
-              <p className="mt-1 text-[13px] leading-relaxed text-navy-3">{senior.reason}</p>
-            )}
-          </div>
-        )}
-
-        {/* Terminal results (GOV-6) — real figures once captured. Each exam is honest-absence-gated on
-            its OWN: NOT_APPLICABLE omitted (omit-not-fake, matching the tier sections above),
-            NOT_CAPTURED shows a reason, CAPTURED shows the figure. Narrows on status → no fabricated
-            number for a not-captured exam. */}
-        {(terminal.bece.status !== "NOT_APPLICABLE" ||
-          terminal.wassce.status !== "NOT_APPLICABLE") && (
-          <div>
-            <div className="text-[10px] font-bold uppercase tracking-[0.12em] text-navy-3">
-              Terminal exams
-            </div>
-            <div className="mt-1 space-y-2">
-              <TerminalLine label="BECE" arm={terminal.bece} />
-              <TerminalLine label="WASSCE" arm={terminal.wassce} />
-            </div>
-          </div>
-        )}
-      </div>
-    </Tile>
-  );
-}
-
-/** One terminal-exam line on the board. NOT_APPLICABLE → omitted; NOT_CAPTURED → reason (no number);
- *  CAPTURED → year · pass rate · passed/total · sex split. Narrowing on `status` makes a fabricated
- *  number for a not-captured exam a compile error. */
-function TerminalLine({ label, arm }: { label: string; arm: RollupArm<TerminalResultSummary> }) {
-  if (arm.status === "NOT_APPLICABLE") return null;
-  return (
-    <div>
-      <div className="text-[11px] font-semibold text-navy-2">{label}</div>
-      {arm.status === "CAPTURED" ? (
-        <div className="mt-0.5 flex flex-wrap items-baseline gap-x-2 gap-y-0.5">
-          <span className="font-display text-xl font-medium text-navy">{arm.data.passRate}%</span>
-          <span className="text-[12px] text-navy-3">pass · {arm.data.year}</span>
-          <span className="text-[12px] text-navy-3">
-            {arm.data.passedCount.toLocaleString("en-GH")}/
-            {arm.data.totalCandidates.toLocaleString("en-GH")} passed
-          </span>
-          <span className="font-mono text-[10px] text-navy-3">
-            {arm.data.female.passed}/{arm.data.female.candidates}F ·{" "}
-            {arm.data.male.passed}/{arm.data.male.candidates}M
-          </span>
-        </div>
-      ) : (
-        <p className="mt-0.5 text-[13px] leading-relaxed text-navy-3">{arm.reason}</p>
-      )}
-    </div>
-  );
-}
-
-/* ───────────────────────────── Infrastructure tile ───────────────────────────── */
-
-/** The board projection of the LATEST facilities snapshot (GOV-7). Narrows on `status`, so a fabricated
- *  number for a not-captured tile is a compile error; a captured zero (0 working / handwashing false)
- *  renders as a real value, never absence. Census-only fields (caterer / furniture / staff FTE) are
- *  structurally absent from `InfrastructureSummary`, so they can never appear here. */
-function InfrastructureTile({ arm }: { arm: RollupArm<InfrastructureSummary> }) {
-  return (
-    <Tile
-      title="Infrastructure"
-      accent="& facilities"
-      meta={
-        arm.status === "CAPTURED"
-          ? `${arm.data.capturedFor.periodLabel} · ${arm.data.capturedFor.academicYear}`
-          : undefined
-      }
-    >
-      {arm.status !== "CAPTURED" ? (
-        <div className="mt-4">
-          <AbsencePanel>{arm.reason}</AbsencePanel>
-        </div>
-      ) : (
-        <InfrastructureBody d={arm.data} />
-      )}
-    </Tile>
-  );
-}
-
-const yesNo = (b: boolean) => (b ? "Yes" : "No");
-const dashNum = (n: number | null) => (n == null ? "—" : n.toLocaleString("en-GH"));
-
-function InfrastructureBody({ d }: { d: InfrastructureSummary }) {
-  return (
-    <div className="mt-3 space-y-4">
-      {/* Classrooms headline — % sound + the count. */}
-      <div className="flex items-center gap-3">
-        <div className="font-display text-3xl font-medium leading-none text-navy">
-          {d.classrooms.pctGood == null ? "—" : `${d.classrooms.pctGood}%`}
-        </div>
-        <div className="text-[11px] leading-tight text-navy-3">
-          classrooms sound
-          <br />
-          {d.classrooms.good}/{d.classrooms.total} good · {d.classrooms.needingRepair} need repair
-        </div>
-      </div>
-
-      {/* Utilities · ICT · library · feeding · textbooks. */}
-      <dl className="space-y-1 text-[13px]">
-        <Line label="Water" value={d.utilities.waterSource} />
-        <Line label="Electricity" value={d.utilities.electricitySource} />
-        <Line
-          label="Sanitation"
-          value={`${d.utilities.latrineType} · ${d.utilities.latrinesTotal.toLocaleString("en-GH")} latrines`}
-        />
-        <Line label="Handwashing" value={yesNo(d.utilities.handwashing)} />
-        <Line
-          label="ICT lab"
-          value={
-            d.ict.hasLab
-              ? `Yes · ${dashNum(d.ict.working)}/${dashNum(d.ict.computers)} working`
-              : "No"
-          }
-        />
-        <Line label="Internet" value={yesNo(d.ict.internet)} />
-        <Line
-          label="Library"
-          value={d.library.has ? `Yes · ${dashNum(d.library.bookCount)} books` : "No"}
-        />
-        <Line
-          label="Feeding"
-          value={
-            d.feeding.gsfpParticipating
-              ? `GSFP · ${dashNum(d.feeding.pupilsFedDaily)} fed daily`
-              : d.feeding.hasKitchen
-                ? "Own kitchen"
-                : "None"
-          }
-        />
-        {d.textbooks.availability && <Line label="Textbooks" value={d.textbooks.availability} />}
-      </dl>
-    </div>
-  );
-}
-
-/* ───────────────────────────── Shared stream/line bits ───────────────────────────── */
-
-const Headline = ({ children }: { children: ReactNode }) => (
-  <div className="mt-2 font-display text-2xl font-medium text-navy">{children}</div>
-);
-const Caption = ({ children }: { children: ReactNode }) => (
-  <div className="mt-0.5 text-[11px] text-navy-3">{children}</div>
-);
-const Reason = ({ children }: { children: ReactNode }) => (
-  <div className="mt-2 text-[13px] leading-relaxed text-navy-3">{children}</div>
-);
